Moved internalportal to separate app
