+++ internalportal/templates/internalportal/internalportal.html
@@ -8,7 +8,6 @@
             <div class="row">
                 <div class="col s12">
                     <h4>{% trans "Internportal" %}</h4>
-                    <p>{{ greeting }}</p>
                 </div>
             </div>
         </div>
