+++ internalportal/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from . import views
+
+app_name = "internalportal"
+
+urlpatterns = [
+    path("", views.InternalPortalView.as_view(), name="internalportal"),
+]
