+++ internalportal/views.py
@@ -1,0 +1,33 @@
+from datetime import datetime
+
+from django.contrib.auth.mixins import PermissionRequiredMixin
+from django.views.generic import TemplateView
+
+from inventory.models.item_loan import ItemLoan
+from news.models import Article, Event
+
+
+class InternalPortalView(PermissionRequiredMixin, TemplateView):
+    template_name = "internalportal/internalportal.html"
+    permission_required = "userprofile.is_active_member"
+
+    def get_context_data(self, *args, **kwargs):
+        context = super().get_context_data(*args, **kwargs)
+
+        context["current_date"] = datetime.now()
+
+        # Find the 5 loan apps that have gone unapproved the longest
+        context["loan_app_list"] = ItemLoan.objects.filter(
+            approver__isnull=True,
+        ).order_by("-loan_from")[:5]
+
+        # Same as in the index view
+        context["event_list"] = Event.objects.filter(internal=True).order_by(
+            "-time_start"
+        )[:5]
+
+        context["article_list"] = Article.objects.filter(
+            internal=True, draft=False
+        ).order_by("-pub_date")[:5]
+
+        return context
