+++ website/settings.py
@@ -65,6 +65,7 @@ INSTALLED_APPS = [
     "social_django",
     "inventory",
     "watchlist",
+    "internalportal",
     "projectarchive",
     "markdownx",
     "django_crontab",
@@ -343,17 +344,6 @@ REST_FRAMEWORK = {
     "DEFAULT_FILTER_BACKENDS": ("django_filters.rest_framework.DjangoFilterBackend",)
 }
 
-# Random greetings that are displayed to the user on the internalportal page
-# Just for fun. The user's first name is formatted in
-INTERNALPORTAL_GREETINGS = [
-    "Hei, {}",  # zzz
-    "Heisann, {}",  # zzzzzz
-    "Halla, {}",  # zzzzzzzzzzzzzzzzz
-    "Og et rungende tjo-bing til deg, {}!",  # real shit?
-    "Husket å skru av ovnen, {}?",
-    "Dette er en tilfeldig melding. Plukket tilfeldig, altså. Selve meldingen er ikke tilfeldig generert ved å drive å plukke bokstaver og sånt. Du skjønner hva jeg mener, {}.",
-    "Lorem ipsum dolor sit amet, consectetur adipiscing elit... Oi, vent, dette er jo i produksjon!",
-]
 #################################
 # Website-specific              #
 #################################
